Crate oak_functions_test_utils

Test utilities to help with unit testing of Oak Functions code.

Structs

Background

Wrapper around a termination signal [oneshot::Sender] and the [JoinHandle] of the associated background task, created by background.

BackgroundHandle

A wrapper around a child process that kills it when its dropped.

Term

A wrapper around a termination signal [oneshot::Receiver].

Statics

MOCK_LOOKUP_DATA_PATH

OAK_RESTRICTED_KERNEL_WRAPPER_BIN

STAGE0

Functions

assert_response_body

background

Executes the provided closure passing to it a Term instance signalling when to terminate, and spawns the resulting Future in the background, returning a Background instance.

create_test_lookup_data

free_port

make_request

run_cc_client

run_java_client

run_oak_functions_containers_example_in_background

Runs the specified example as a background task. Returns a reference to the running server and the port on which the server is listening.

run_oak_functions_example_in_background

Runs the specified example as a background task. Returns a reference to the running server and the port on which the server is listening.

rust_crate_enclave_out_path

rust_crate_wasm_out_path

serialize_entries

Serializes the provided map as a contiguous buffer of length-delimited protobuf messages of type Entry.

skip_test

Whether to skip the test. For instance, GitHub Actions does not support KVM, so we cannot run tests that require nested virtualization.

workspace_path

write_to_temp_file