Crate oak_functions_test_utils

Test utilities to help with unit testing of Oak Functions code.

Structs

Background

Wrapper around a termination signal [oneshot::Sender] and the [JoinHandle] of the associated background task, created by background.

Term

A wrapper around a termination signal [oneshot::Receiver].

Statics

MOCK_LOOKUP_DATA_PATH

OAK_RESTRICTED_KERNEL_WRAPPER_BIN

Functions

assert_response_body

background

Executes the provided closure passing to it a Term instance signalling when to terminate, and spawns the resulting Future in the background, returning a Background instance.

build_rust_crate_enclave

build_rust_crate_linux

Builds the crate identified by the given package name (as per the name attribute in a Cargo.toml file included in the root cargo workspace) as a Linux binary, and returns the path of the resulting binary.

build_rust_crate_wasm

Builds the crate identified by the given package name (as per the name attribute in a Cargo.toml file included in the root cargo workspace) as a Wasm module, and returns the path of the resulting binary.

create_test_lookup_data

free_port

kill_process

Kills all the processes identified by the provided handle.

make_request

run_oak_functions_containers_example_in_background

Runs the specified example as a background task. Returns a reference to the running server and the port on which the server is listening.

run_oak_functions_example_in_background

Runs the specified example as a background task. Returns a reference to the running server and the port on which the server is listening.

rust_crate_wasm_out_path

Returns the path of the Wasm module generated by the crate identified by the given package name.

serialize_entries

Serializes the provided map as a contiguous buffer of length-delimited protobuf messages of type Entry.

workspace_path

write_to_temp_file