Function oak_functions_test_utils::background

pub fn background<Out, F>(f: F) -> Background<Out::Output>
where
    Out: Future + Send + 'static,
    Out::Output: Send + 'static,
    F: FnOnce(Term) -> Out,

Executes the provided closure passing to it a Term instance signalling when to terminate, and spawns the resulting Future in the background, returning a Background instance.